added validation to UserController and OrderController use spring validation. (without exception handling). Changed application.properties from ddl-auto=none to ddl-auto=validate.

+++ src/main/java/com/hordiienko/onlinestore/entity/Role.java
@@ -14,17 +14,17 @@ import java.util.Set;
 @NoArgsConstructor
 public class Role implements GrantedAuthority {
     @Id
-    private Long id;
+    private Integer id;
     private String name;
     @Transient
     @ManyToMany(mappedBy = "roles")
     private Set<User> users;
 
-    public Role(Long id) {
+    public Role(Integer id) {
         this.id = id;
     }
 
-    public Role(Long id, String name) {
+    public Role(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
